Conflicts:
	nbproject/project.properties

+++ src/java/service/AdminService.java
@@ -5,12 +5,17 @@
  */
 package service;
 
+import dao.CabinetUserDao;
 import dao.PersonalCabinetDao;
+import dao.StrategyDao;
+import dao.TagDao;
 import dao.TarifDao;
 import entities.CabinetUser;
 import entities.Campaign;
 import entities.Client;
 import entities.PersonalCabinet;
+import entities.Strategy;
+import entities.Tag;
 import entities.Tarif;
 import entities.User;
 import java.util.Calendar;
@@ -34,6 +39,15 @@ public class AdminService extends PrimService {
 
     @Autowired
     PersonalCabinetDao pkDao;
+
+    @Autowired
+    CabinetUserDao cuDao;
+
+    @Autowired
+    StrategyDao stratDao;
+
+    @Autowired
+    TagDao tagDao;
 
     @Autowired
     TarifDao tarifDao;
@@ -122,63 +136,13 @@ public class AdminService extends PrimService {
                 PersonalCabinet pk = pkDao.find(pkId);
                 if (pk != null) {
                     Tarif t = pk.getTarif();
-                    if(t!=null){
+                    if (t != null) {
                         Long uc = t.getUserCount();
                         List<CabinetUser> culist = pkDao.getActiveUserList(pkId);
-                        if(uc==null||culist.size()<uc.intValue()){
+                        if (uc == null || culist.size() < uc.intValue()) {
                             return true;
                         }
-                    }else{
-                        return true;
-                    }
-                } else {
-                    addError("Личный кабинет с ид " + pkId + " не был найден");
-                }
-            } else {
-                addError("Ид личного кабинета не передан");
-            }
-        }
-        return false;
-    }
-    
-    public boolean mayAddCampaign(Long pkId) {
-        if (tarifIsNotExpired(pkId)) {
-            if (pkId != null) {
-                PersonalCabinet pk = pkDao.find(pkId);
-                if (pk != null) {
-                    Tarif t = pk.getTarif();
-                    if(t!=null){
-                        Long uc = t.getCampaignCount();
-                        List<Campaign> clist = pkDao.getCampaignList(pkId);
-                        if(uc==null||clist.size()<uc.intValue()){
-                            return true;
-                        }
-                    }else{
-                        return true;
-                    }
-                } else {
-                    addError("Личный кабинет с ид " + pkId + " не был найден");
-                }
-            } else {
-                addError("Ид личного кабинета не передан");
-            }
-        }
-        return false;
-    }
-    
-    public boolean mayAddClient(Long pkId) {
-        if (tarifIsNotExpired(pkId)) {
-            if (pkId != null) {
-                PersonalCabinet pk = pkDao.find(pkId);
-                if (pk != null) {
-                    Tarif t = pk.getTarif();
-                    if(t!=null){
-                        Long uc = t.getClientCount();
-                        List<Client> clist = pkDao.getClientList(pkId);
-                        if(uc==null||clist.size()<uc.intValue()){
-                            return true;
-                        }
-                    }else{
+                    } else {
                         return true;
                     }
                 } else {
@@ -191,9 +155,77 @@ public class AdminService extends PrimService {
         return false;
     }
 
-     public void delete(Long pkId) {
-         PersonalCabinet pk = pkDao.find(pkId);
-         pkDao.delete(pk);
-     }
-    
+    public boolean mayAddCampaign(Long pkId) {
+        if (tarifIsNotExpired(pkId)) {
+            if (pkId != null) {
+                PersonalCabinet pk = pkDao.find(pkId);
+                if (pk != null) {
+                    Tarif t = pk.getTarif();
+                    if (t != null) {
+                        Long uc = t.getCampaignCount();
+                        List<Campaign> clist = pkDao.getCampaignList(pkId);
+                        if (uc == null || clist.size() < uc.intValue()) {
+                            return true;
+                        }
+                    } else {
+                        return true;
+                    }
+                } else {
+                    addError("Личный кабинет с ид " + pkId + " не был найден");
+                }
+            } else {
+                addError("Ид личного кабинета не передан");
+            }
+        }
+        return false;
+    }
+
+    public boolean mayAddClient(Long pkId) {
+        if (tarifIsNotExpired(pkId)) {
+            if (pkId != null) {
+                PersonalCabinet pk = pkDao.find(pkId);
+                if (pk != null) {
+                    Tarif t = pk.getTarif();
+                    if (t != null) {
+                        Long uc = t.getClientCount();
+                        List<Client> clist = pkDao.getClientList(pkId);
+                        if (uc == null || clist.size() < uc.intValue()) {
+                            return true;
+                        }
+                    } else {
+                        return true;
+                    }
+                } else {
+                    addError("Личный кабинет с ид " + pkId + " не был найден");
+                }
+            } else {
+                addError("Ид личного кабинета не передан");
+            }
+        }
+        return false;
+    }
+
+    public void delete(Long pkId) {
+        PersonalCabinet pk = pkDao.find(pkId);
+
+        if (pk != null) {
+            List<Strategy> strList = pk.getStrategyList();
+            for (Strategy str : strList) {
+                stratDao.delete(str);
+            }
+
+            List<Tag> tglist = tagDao.getAllTags(pkId);
+            for (Tag tg : tglist) {
+                tagDao.delete(tg);
+            }
+
+            List<CabinetUser> lku = pk.getCabinetUserList();
+            for (CabinetUser cu : lku) {
+                cuDao.delete(cu);
+            }
+            pkDao.delete(pk);
+        }
+        
+    }
+
 }
